Delete package_rakwireless_com_rui_index.json
Because this file is nearly identical to `package_rakwireless.com_rui_index.json`. The only difference is the number of tab spaces before the `{` on line 1577, which is purely cosmetic anyway. 

Plus, this [code change](https://github.com/RAKWireless/RAKwireless-Arduino-BSP-Index/issues/7) indicates the new name is the correct one. Not sure personally if this is the case, but regardless, there is no need for both files as it causes confusion on which one to use.